Move copy query button to right bottom of sql editor box
